Created Relationship between Bootcamp Courses and User and updated the required routes

+++ controllers/coursesController.js
@@ -43,11 +43,16 @@ exports.getCourse = asyncHandler ( async (req, res, next) =>{
 exports.addCourse = asyncHandler ( async (req, res, next) =>{
 
     req.body.bootcamp = req.params.bootcampId
+    req.body.user = req.user.id;
    
     const singleBootcamp = await bootcamp.findById(req.params.bootcampId);
 
     if(!singleBootcamp) {
         return next(new ErrorResponse(`No Bootcamp with the id of ${req.params.bootcampId}`, 404));
+    }
+
+    if(singleBootcamp.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to add a course to ${singleBootcamp._id}`, 401));
     }
 
     const newCourse = await course.create(req.body);
@@ -67,6 +72,11 @@ exports.updateCourse = asyncHandler ( async (req, res, next) =>{
         return next(new ErrorResponse(`No course with the id of ${req.params.id}`, 404));
     }
 
+    if(updatedCourse.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to update course ${course._id}`, 401));
+    }
+
+
     updatedCourse = await course.findByIdAndUpdate(req.params.id, req.body, {
         new : true,
         runValidators : true
@@ -84,6 +94,10 @@ exports.deleteCourse = asyncHandler ( async (req, res, next) =>{
 
     if(!deletedCourse) {
         return next(new ErrorResponse(`No course with the id of ${req.params.id}`, 404));
+    }
+
+    if(deletedCourse.user.toString() !== req.user.id && req.user.role !== 'admin'){
+        return next(new ErrorResponse(`User ${req.user.id} is not authorized to delete course ${course._id}`, 401));
     }
 
     await deletedCourse.deleteOne();
